requirements.txt
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
opencv-contrib-python==4.4.0.44
pandas==1.2.4
Pillow==8.2.0
pykitti==0.3.1
PyOpenGL @ file:///C:/Users/henri/Documents/ComputerVision/kitti_visualization_pyqt/prerequisites/PyOpenGL-3.1.5-cp38-cp38-win_amd64.whl
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
